Consolidated UI for all profile ops.
