+++ src/Mod/Path/Gui/Resources/Path.qrc
@@ -64,10 +64,8 @@
         <file>panels/PageBaseGeometryEdit.ui</file>
         <file>panels/PageDepthsEdit.ui</file>
         <file>panels/PageHeightsEdit.ui</file>
-        <file>panels/PageOpContourEdit.ui</file>
         <file>panels/PageOpPocketEdit.ui</file>
-        <file>panels/PageOpProfileEdgesEdit.ui</file>
-        <file>panels/PageOpProfileFacesEdit.ui</file>
+        <file>panels/PageOpProfileFullEdit.ui</file>
         <file>panels/PocketEdit.ui</file>
         <file>panels/PointEdit.ui</file>
         <file>panels/ProfileEdgesEdit.ui</file>
