+++ src/Mod/Path/Gui/Resources/panels/PageOpProfileFullEdit.ui
@@ -40,7 +40,7 @@
     <widget class="QWidget" name="widget" native="true">
      <layout class="QGridLayout" name="gridLayout">
       <item row="0" column="0">
-       <widget class="QLabel" name="label_2">
+       <widget class="QLabel" name="cutSideLabel">
         <property name="text">
          <string>Cut Side</string>
         </property>
@@ -61,7 +61,7 @@
        </widget>
       </item>
       <item row="1" column="0">
-       <widget class="QLabel" name="label_3">
+       <widget class="QLabel" name="directionLabel">
         <property name="text">
          <string>Direction</string>
         </property>
@@ -82,7 +82,7 @@
        </widget>
       </item>
       <item row="2" column="0">
-       <widget class="QLabel" name="label_4">
+       <widget class="QLabel" name="extraOffsetLabel">
         <property name="text">
          <string>Extra Offset</string>
         </property>
