+++ src/Mod/Path/PathScripts/PathProfileFacesGui.py
@@ -23,60 +23,16 @@
 # ***************************************************************************
 
 import FreeCAD
-import FreeCADGui
-import Path
 import PathScripts.PathAreaOpGui as PathAreaOpGui
-import PathScripts.PathLog as PathLog
+import PathScripts.PathProfileBaseGui as PathProfileBaseGui
 import PathScripts.PathProfileFaces as PathProfileFaces
-import PathScripts.PathSelection as PathSelection
 
-from PathScripts import PathUtils
-from PySide import QtCore, QtGui
+from PySide import QtCore
 
-def translate(context, text, disambig=None):
-    return QtCore.QCoreApplication.translate(context, text, disambig)
+class TaskPanelOpPage(PathProfileBaseGui.TaskPanelOpPage):
 
-class TaskPanelOpPage(PathAreaOpGui.TaskPanelPage):
-
-    def getForm(self):
-        return FreeCADGui.PySideUic.loadUi(":/panels/PageOpProfileFacesEdit.ui")
-
-    def getFields(self, obj):
-        self.obj.OffsetExtra = FreeCAD.Units.Quantity(self.form.extraOffset.text()).Value
-        self.obj.UseComp = self.form.useCompensation.isChecked()
-        self.obj.UseStartPoint = self.form.useStartPoint.isChecked()
-        self.obj.Side = str(self.form.cutSide.currentText())
-        self.obj.Direction = str(self.form.direction.currentText())
-        self.obj.processHoles = self.form.processHoles.isChecked()
-        self.obj.processPerimeter = self.form.processPerimeter.isChecked()
-        self.obj.processCircles = self.form.processCircles.isChecked()
-
-        tc = PathUtils.findToolController(self.obj, self.form.toolController.currentText())
-        self.obj.ToolController = tc
-
-    def setFields(self, obj):
-        self.form.extraOffset.setText(FreeCAD.Units.Quantity(self.obj.OffsetExtra.Value, FreeCAD.Units.Length).UserString)
-        self.form.useCompensation.setChecked(self.obj.UseComp)
-        self.form.useStartPoint.setChecked(self.obj.UseStartPoint)
-        self.form.processHoles.setChecked(self.obj.processHoles)
-        self.form.processPerimeter.setChecked(self.obj.processPerimeter)
-        self.form.processCircles.setChecked(self.obj.processCircles)
-
-        self.selectInComboBox(self.obj.Side, self.form.cutSide)
-        self.selectInComboBox(self.obj.Direction, self.form.direction)
-        self.setupToolController(self.obj, self.form.toolController)
-
-    def getSignalsForUpdate(self, obj):
-        signals = []
-        signals.append(self.form.cutSide.currentIndexChanged)
-        signals.append(self.form.direction.currentIndexChanged)
-        signals.append(self.form.useCompensation.clicked)
-        signals.append(self.form.useStartPoint.clicked)
-        signals.append(self.form.extraOffset.editingFinished)
-        signals.append(self.form.processHoles.clicked)
-        signals.append(self.form.processPerimeter.clicked)
-        signals.append(self.form.processCircles.clicked)
-        return signals
+    def profileFeatures(self):
+        return PathProfileBaseGui.FeatureSide | PathProfileBaseGui.FeatureProcessing
 
 PathAreaOpGui.SetupOperation('Profile Faces',
         PathProfileFaces.Create,
